+++ java/connectors/okx/src/main/java/com/epam/deltix/data/connectors/okx/OkxSpotConnectorSettings.java
@@ -1,18 +1,18 @@
-package com.epam.deltix.data.connectors.okex;
+package com.epam.deltix.data.connectors.okx;
 
 import com.epam.deltix.data.connectors.commons.DataConnectorSettings;
 import com.epam.deltix.data.connectors.commons.annotations.ConnectorSettings;
 
-@ConnectorSettings("OKEX")
-public class OkexSpotConnectorSettings extends DataConnectorSettings {
+@ConnectorSettings("OKX")
+public class OkxSpotConnectorSettings extends DataConnectorSettings {
 
     private String wsUrl = "wss://ws.okex.com:8443/ws/v5/public";
 
-    public OkexSpotConnectorSettings() {
+    public OkxSpotConnectorSettings() {
         super();
     }
 
-    public OkexSpotConnectorSettings(String tbUrl, String stream, String wsUrl) {
+    public OkxSpotConnectorSettings(String tbUrl, String stream, String wsUrl) {
         super(tbUrl, stream);
         this.wsUrl = wsUrl;
     }
